Move: test util file 이동
